+++ trunk/engine/toolkit/contentbrowser/code/userinterface.cpp
@@ -174,7 +174,7 @@ namespace Toolkit
 
 	void UserInterface::RenderDocks()
 	{		
-		ImGui::RootDock(ImVec2(0, 16), ImVec2(application->window->GetWidth(), application->window->GetHeight() - 16));
+		ImGui::RootDock(ImVec2(0.0f, 16.0f), ImVec2((float)application->window->GetWidth(), (float)application->window->GetHeight() - 16.0f));
 		{
 			ImGui::BeginDock("3D View", NULL, ImGuiWindowFlags_NoSavedSettings | ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_NoScrollWithMouse);
 			{
